Refactor repository structure (#322)
Change Summary
- Move out extractor, descheduler, syncer, kpis into separate service
"boxes", rename internal/ to lib/
- Move out shared objects to api subfolders
- Remove outdated monoservice setup including associated workflows,
Dockerfile, Makefile, etc.
- Rename cortex-core chart to cortex-alerts (since it now only contains
alerts)
- Refactor Tiltfile to support new setup
- Change how indexes are defined to not have to import lib module in api

---------

+++ Tiltfile
@@ -35,18 +35,54 @@ def kubebuilder_binary_files(path):
 
 ########### Cortex Scheduler
 docker_build('ghcr.io/cobaltcore-dev/cortex-scheduler', '.',
-    dockerfile='Dockerfile.kubebuilder',
+    dockerfile='Dockerfile',
     build_args={'GO_MOD_PATH': 'scheduler'},
-    only=kubebuilder_binary_files('scheduler') + ['reservations/', 'decisions/', 'internal/', 'go.mod', 'go.sum'],
+    only=kubebuilder_binary_files('scheduler') + ['reservations/', 'decisions/', 'extractor/', 'sync/', 'lib/', 'testlib/'],
 )
 local('sh helm/sync.sh scheduler/dist/chart')
 # Deployed as part of bundles below.
 
+########### Cortex Descheduler
+docker_build('ghcr.io/cobaltcore-dev/cortex-descheduler', '.',
+    dockerfile='Dockerfile',
+    build_args={'GO_MOD_PATH': 'descheduler'},
+    only=kubebuilder_binary_files('descheduler') + ['lib/', 'testlib/', 'sync/'],
+)
+local('sh helm/sync.sh descheduler/dist/chart')
+# Deployed as part of bundles below.
+
+########### Cortex Extractor
+docker_build('ghcr.io/cobaltcore-dev/cortex-extractor', '.',
+    dockerfile='Dockerfile',
+    build_args={'GO_MOD_PATH': 'extractor'},
+    only=kubebuilder_binary_files('extractor') + ['lib/', 'testlib/', 'sync/'],
+)
+local('sh helm/sync.sh extractor/dist/chart')
+# Deployed as part of bundles below.
+
+########### Cortex KPIs
+docker_build('ghcr.io/cobaltcore-dev/cortex-kpis', '.',
+    dockerfile='Dockerfile',
+    build_args={'GO_MOD_PATH': 'kpis'},
+    only=kubebuilder_binary_files('kpis') + ['lib/', 'testlib/', 'sync/', 'extractor/'],
+)
+local('sh helm/sync.sh kpis/dist/chart')
+# Deployed as part of bundles below.
+
+########### Cortex Syncer
+docker_build('ghcr.io/cobaltcore-dev/cortex-syncer', '.',
+    dockerfile='Dockerfile',
+    build_args={'GO_MOD_PATH': 'sync'},
+    only=kubebuilder_binary_files('sync') + ['lib/', 'testlib/'],
+)
+local('sh helm/sync.sh sync/dist/chart')
+# Deployed as part of bundles below.
+
 ########### Reservations Operator & CRDs
 docker_build('ghcr.io/cobaltcore-dev/cortex-reservations-operator', '.',
-    dockerfile='Dockerfile.kubebuilder',
+    dockerfile='Dockerfile',
     build_args={'GO_MOD_PATH': 'reservations'},
-    only=kubebuilder_binary_files('reservations') + ['scheduler/', 'decisions/', 'internal/', 'go.mod', 'go.sum'],
+    only=kubebuilder_binary_files('reservations') + ['scheduler/', 'decisions/', 'lib/', 'testlib/'],
 )
 local('sh helm/sync.sh reservations/dist/chart')
 k8s_yaml(helm('reservations/dist/chart', name='cortex-reservations', values=[tilt_values]))
@@ -54,52 +90,64 @@ k8s_resource('reservations-controller-manager', labels=['Reservations'])
 
 ########### Decisions Operator & CRDs
 docker_build('ghcr.io/cobaltcore-dev/cortex-decisions-operator', '.',
-    dockerfile='Dockerfile.kubebuilder',
+    dockerfile='Dockerfile',
     build_args={'GO_MOD_PATH': 'decisions'},
-    only=kubebuilder_binary_files('decisions') + ['internal/', 'go.mod', 'go.sum'],
+    only=kubebuilder_binary_files('decisions') + ['lib/', 'testlib/'],
 )
 local('sh helm/sync.sh decisions/dist/chart')
 k8s_yaml(helm('decisions/dist/chart', name='cortex-decisions', values=[tilt_values]))
 k8s_resource('decisions-controller-manager', labels=['Decisions'])
 
 ########### Cortex Bundles
-docker_build('ghcr.io/cobaltcore-dev/cortex', '.', only=[
-    'internal/', 'commands/', 'main.go', 'go.mod', 'go.sum', 'Makefile',
-    'reservations/api/', # API module of the reservations operator needed for the scheduler.
-    'decisions/api/', # API module of the decisions operator needed for the scheduler.
-])
 docker_build('ghcr.io/cobaltcore-dev/cortex-postgres', 'postgres')
 
 # Package the lib charts locally and sync them to the bundle charts. In this way
 # we can bump the lib charts locally and test them before pushing them to the OCI registry.
 
-dep_charts = [
-    ('helm/library/cortex-core', 'cortex-core'),
+dep_charts = lambda bundle_chart_name: [
+    ('helm/library/cortex-alerts', 'cortex-alerts'),
     ('helm/library/cortex-postgres', 'cortex-postgres'),
     ('helm/library/cortex-mqtt', 'cortex-mqtt'),
+
     ('scheduler/dist/chart', 'cortex-scheduler'),
-]
+    ('extractor/dist/chart', 'cortex-extractor'),
+    ('kpis/dist/chart', 'cortex-kpis'),
+    ('sync/dist/chart', 'cortex-syncer'),
+] + ([
+    ('descheduler/dist/chart', 'cortex-descheduler'),
+] if bundle_chart_name in ['cortex-nova'] else [])
+
 bundle_charts = [
     ('helm/bundles/cortex-nova', 'cortex-nova'),
     ('helm/bundles/cortex-manila', 'cortex-manila'),
     ('helm/bundles/cortex-cinder', 'cortex-cinder'),
 ]
 
-for (dep_chart_path, dep_chart_name) in dep_charts:
-    watch_file(dep_chart_path)
-    local('sh helm/sync.sh ' + dep_chart_path)
-    for (bundle_chart_path, bundle_chart_name) in bundle_charts:
+for (bundle_chart_path, bundle_chart_name) in bundle_charts:
+    for (dep_chart_path, dep_chart_name) in dep_charts(bundle_chart_name):
+        print('--- Syncing dependency ' + dep_chart_name + ' into bundle ' + bundle_chart_name)
+        watch_file(dep_chart_path)
+        local('sh helm/sync.sh ' + dep_chart_path)
         local('helm package ' + dep_chart_path)
         gen_tgz = str(local('ls ' + dep_chart_name + '-*.tgz')).strip()
+        # If the file isn't there yet, copy it over.
+        if not os.path.exists(bundle_chart_path + '/charts/' + gen_tgz):
+            print('Adding ' + dep_chart_name + ' to ' + bundle_chart_name)
+            local('mkdir -p helm/bundles/' + bundle_chart_name + '/charts/')
+            local('mv -f ' + gen_tgz + ' ' + bundle_chart_path + '/charts/')
+            continue
+        # If it is there, compare the files and only copy if they differ.
         cmp = 'sh helm/cmp.sh ' + gen_tgz + ' ' + bundle_chart_path + '/charts/' + gen_tgz
         cmp_result = str(local(cmp)).strip()
         if cmp_result == 'true':
             print('Skipping ' + dep_chart_name + ' as it is already up to date in ' + bundle_chart_name)
             local('rm -f ' + gen_tgz)
         else:
+            print('Updating ' + dep_chart_name + ' in ' + bundle_chart_name)
             local('mkdir -p helm/bundles/' + bundle_chart_name + '/charts/')
             local('mv -f ' + gen_tgz + ' ' + bundle_chart_path + '/charts/')
 for (bundle_chart_path, _) in bundle_charts:
+    print('--- Final sync of bundle chart: ' + bundle_chart_path)
     local('sh helm/sync.sh ' + bundle_chart_path)
 
 port_mappings = {}
@@ -110,40 +158,43 @@ def new_port_mapping(component, local_port, remote_port):
 if 'nova' in ACTIVE_DEPLOYMENTS:
     print("Activating Cortex Nova bundle")
     k8s_yaml(helm('./helm/bundles/cortex-nova', name='cortex-nova', values=[tilt_values]))
-    k8s_resource('cortex-nova-postgresql', labels=['Cortex-Nova'])
+    k8s_resource('cortex-nova-postgresql', labels=['Cortex-Nova'], port_forwards=[
+        new_port_mapping('cortex-nova-postgresql', 8000, 5432),
+    ])
     k8s_resource('cortex-nova-mqtt', labels=['Cortex-Nova'])
-    k8s_resource('cortex-nova-migrations', labels=['Cortex-Nova'])
     k8s_resource('cortex-nova-syncer', labels=['Cortex-Nova'])
     k8s_resource('cortex-nova-extractor', labels=['Cortex-Nova'])
     k8s_resource('cortex-nova-kpis', labels=['Cortex-Nova'])
     k8s_resource('cortex-nova-descheduler', labels=['Cortex-Nova'])
     k8s_resource('cortex-nova-scheduler', labels=['Cortex-Nova'], port_forwards=[
-        new_port_mapping('cortex-nova-scheduler-api', 8000, 8080),
+        new_port_mapping('cortex-nova-scheduler-api', 8001, 8080),
     ])
 
 if 'manila' in ACTIVE_DEPLOYMENTS:
     print("Activating Cortex Manila bundle")
     k8s_yaml(helm('./helm/bundles/cortex-manila', name='cortex-manila', values=[tilt_values]))
-    k8s_resource('cortex-manila-postgresql', labels=['Cortex-Manila'])
+    k8s_resource('cortex-manila-postgresql', labels=['Cortex-Manila'], port_forwards=[
+        new_port_mapping('cortex-manila-postgresql', 8002, 5432),
+    ])
     k8s_resource('cortex-manila-mqtt', labels=['Cortex-Manila'])
-    k8s_resource('cortex-manila-migrations', labels=['Cortex-Manila'])
     k8s_resource('cortex-manila-syncer', labels=['Cortex-Manila'])
     k8s_resource('cortex-manila-extractor', labels=['Cortex-Manila'])
     k8s_resource('cortex-manila-kpis', labels=['Cortex-Manila'])
     k8s_resource('cortex-manila-scheduler', labels=['Cortex-Manila'], port_forwards=[
-        new_port_mapping('cortex-manila-scheduler-api', 8001, 8080),
+        new_port_mapping('cortex-manila-scheduler-api', 8003, 8080),
     ])
 
 if 'cinder' in ACTIVE_DEPLOYMENTS:
     k8s_yaml(helm('./helm/bundles/cortex-cinder', name='cortex-cinder', values=[tilt_values]))
-    k8s_resource('cortex-cinder-postgresql', labels=['Cortex-Cinder'])
+    k8s_resource('cortex-cinder-postgresql', labels=['Cortex-Cinder'], port_forwards=[
+        new_port_mapping('cortex-cinder-postgresql', 8004, 5432),
+    ])
     k8s_resource('cortex-cinder-mqtt', labels=['Cortex-Cinder'])
-    k8s_resource('cortex-cinder-migrations', labels=['Cortex-Cinder'])
     k8s_resource('cortex-cinder-syncer', labels=['Cortex-Cinder'])
     k8s_resource('cortex-cinder-extractor', labels=['Cortex-Cinder'])
     k8s_resource('cortex-cinder-kpis', labels=['Cortex-Cinder'])
     k8s_resource('cortex-cinder-scheduler', labels=['Cortex-Cinder'], port_forwards=[
-        new_port_mapping('cortex-cinder-scheduler-api', 8002, 8080),
+        new_port_mapping('cortex-cinder-scheduler-api', 8005, 8080),
     ])
 
 ########### Dev Dependencies
